+++ src/Controller/Component/AuthComponent.php
@@ -11,27 +11,39 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/me-cms
  * @license     https://opensource.org/licenses/mit-license.php MIT License
- * @see         http://api.cakephp.org/3.7/class-Cake.Controller.Component.AuthComponent.html
  */
 
 namespace MeCms\Controller\Component;
 
 use Cake\Controller\Component\AuthComponent as CakeAuthComponent;
+use Cake\Controller\ComponentRegistry;
 use MeCms\AuthTrait;
 
 /**
  * Authentication control component class.
  *
  * Binds access control with user authentication and session management.
+ * @deprecated 2.31.0 Use instead the `AuthenticationComponent`
  */
 class AuthComponent extends CakeAuthComponent
 {
     use AuthTrait;
 
     /**
+     * Constructor
+     * @param \Cake\Controller\ComponentRegistry $registry A component registry
+     * @param array<string, mixed> $config Array of configuration settings
+     */
+    public function __construct(ComponentRegistry $registry, array $config = [])
+    {
+        deprecationWarning('Deprecated. Use instead the `AuthenticationComponent`');
+
+        parent::__construct($registry, $config);
+    }
+
+    /**
      * Constructor hook method
-     * @param array $config The configuration settings provided to this
-     *  component
+     * @param array $config The configuration settings provided to this component
      * @return void
      */
     public function initialize(array $config): void