+++ tests/TestCase/Controller/Component/AuthComponentTest.php
@@ -24,12 +24,39 @@ use MeTools\TestSuite\ComponentTestCase;
 class AuthComponentTest extends ComponentTestCase
 {
     /**
+     * @var int
+     */
+    protected static int $initialErrorReporting;
+
+    /**
      * @var array<string>
      */
     public $fixtures = [
         'plugin.MeCms.Users',
         'plugin.MeCms.UsersGroups',
     ];
+
+    /**
+     * This method is called before the first test of this test class is run
+     * @return void
+     */
+    public static function setUpBeforeClass(): void
+    {
+        parent::setUpBeforeClass();
+
+        self::$initialErrorReporting = error_reporting(E_ALL & ~E_USER_DEPRECATED);
+    }
+
+    /**
+     * This method is called after the last test of this test class is run
+     * @return void
+     */
+    public static function tearDownAfterClass(): void
+    {
+        parent::tearDownAfterClass();
+
+        error_reporting(self::$initialErrorReporting);
+    }
 
     /**
      * Tests for `initialize()` method